+++ packages/esm-patient-orders-app/src/order-basket/general-order-type/general-order-panel.scss
