+++ packages/esm-patient-orders-app/src/order-basket/general-order-type/order-basket-item-tile.scss
