+++ packages/esm-patient-orders-app/src/order-basket/general-order-type/orderable-concept-search/search-results.scss
